web/requirements.txt
flask>=3.1
flask-login>=0.6
flask-sqlalchemy>=3.1
playwright>=1.53
anthropic>=0.52
python-dotenv>=1.0
apscheduler>=3.10
cryptography>=44.0
requests>=2.32
